Seki
Fixed_FadeManager.h
Add_Comment

+++ Scripts/Engine/Scene/Calendar/Month.h
@@ -2,6 +2,7 @@
  *	@file	Month.h
  *	@author	Seki
  */
+
 #ifndef _MONTH_H_
 #define _MONTH_H_
 
@@ -13,9 +14,12 @@
  */
 class Month : public DateBase {
 public:
-    std::vector<std::shared_ptr<Week>> weeks;
-    size_t currentWeekIndex = 0;
+    std::vector<std::shared_ptr<Week>> weeks;   // Week配列
+    int currentWeekIndex = 0;                   // 週間カウント
 
+    /*
+     *  進行処理
+     */
     void Advance() override {
         auto week = GetCurrentWeek();
         if (!week) return;
@@ -27,12 +31,16 @@ public:
             weeks.erase(weeks.begin() + currentWeekIndex);
             // currentWeekIndex はそのまま、次の Week が現在のインデックスに来る
         }
-
+        // 進行処理コールバック
         if (onAdvance) onAdvance();
     }
-
+    /*
+     *  一週間の行動終了フラグ
+     */
     bool IsFinished() const override { return weeks.empty(); }
-
+    /*
+     *  現在稼働しているWeekの取得
+     */
     std::shared_ptr<Week> GetCurrentWeek() {
         if (currentWeekIndex < weeks.size()) return weeks[currentWeekIndex];
         return nullptr;
